metamaterials-lab/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
python-multipart
python-dotenv
matplotlib
scipy